move functions with their structs
